# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AssessmentDetail": {
        "description": "Summary plus the full engine payload.\n\n`result` is intentionally untyped here. It is the engine's own contract and\nmirroring its 200-odd fields in Pydantic would create a second definition\nthat drifts. The engine's shape is documented in docs/14-API-REFERENCE.md\nand pinned by the engine tests.",
        "properties": {
          "cash_positive_abatement_tco2e": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cash Positive Abatement Tco2E"
          },
          "cash_positive_benefit_inr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cash Positive Benefit Inr"
          },
          "cash_positive_capex_inr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cash Positive Capex Inr"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "critical_leak_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Critical Leak Count"
          },
          "data_quality_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Quality Score"
          },
          "engine_profile": {
            "additionalProperties": true,
            "title": "Engine Profile",
            "type": "object"
          },
          "factory_id": {
            "title": "Factory Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_baseline": {
            "title": "Is Baseline",
            "type": "boolean"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "leak_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leak Count"
          },
          "result": {
            "additionalProperties": true,
            "title": "Result",
            "type": "object"
          },
          "scope1_tco2e": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope1 Tco2E"
          },
          "scope2_tco2e": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope2 Tco2E"
          },
          "scope3_tco2e": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope3 Tco2E"
          },
          "total_high_tco2e": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total High Tco2E"
          },
          "total_low_tco2e": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Low Tco2E"
          },
          "total_tco2e": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Tco2E"
          },
          "version_stamp": {
            "additionalProperties": true,
            "title": "Version Stamp",
            "type": "object"
          }
        },
        "required": [
          "id",
          "factory_id",
          "is_baseline",
          "created_at",
          "engine_profile",
          "result"
        ],
        "title": "AssessmentDetail",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlantProfileIn": {
        "description": "Exactly what `engine.assess()` consumes. Field names are engine-defined.",
        "properties": {
          "annual_output_t": {
            "default": 0,
            "minimum": 0.0,
            "title": "Annual Output T",
            "type": "number"
          },
          "annual_revenue_cr": {
            "default": 0,
            "minimum": 0.0,
            "title": "Annual Revenue Cr",
            "type": "number"
          },
          "discount_rate": {
            "anyOf": [
              {
                "exclusiveMaximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discount Rate"
          },
          "electricity_kwh": {
            "default": 0,
            "minimum": 0.0,
            "title": "Electricity Kwh",
            "type": "number"
          },
          "employees": {
            "default": 0,
            "minimum": 0.0,
            "title": "Employees",
            "type": "integer"
          },
          "eu_export_share_pct": {
            "default": 0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Eu Export Share Pct",
            "type": "number"
          },
          "freight": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Freight",
            "type": "object"
          },
          "fuels": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Fuels",
            "type": "object"
          },
          "materials": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Materials",
            "type": "object"
          },
          "name": {
            "default": "Unnamed plant",
            "title": "Name",
            "type": "string"
          },
          "sector": {
            "title": "Sector",
            "type": "string"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "tariff_inr_per_kwh": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tariff Inr Per Kwh"
          },
          "waste": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Waste",
            "type": "object"
          }
        },
        "required": [
          "sector"
        ],
        "title": "PlantProfileIn",
        "type": "object"
      },
...
      "RunAssessmentRequest": {
        "properties": {
          "label": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          }
        },
        "title": "RunAssessmentRequest",
        "type": "object"
      },
...
      "ScenarioOut": {
        "properties": {
          "baseline_assessment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Assessment Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "factory_id": {
            "title": "Factory Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "latest_assessment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Assessment Id"
          },
          "modifications": {
            "additionalProperties": true,
            "title": "Modifications",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "factory_id",
          "name",
          "modifications",
          "created_at"
        ],
        "title": "ScenarioOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Industrial Carbon Intelligence Network. Measure, detect, decide, connect, implement, verify. Screening and decision support - not an accredited audit, legal assurance service or carbon-credit verification.",
    "title": "PRANGARA",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/assess": {
      "post": {
        "operationId": "sandbox_assess_api_assess_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlantProfileIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Sandbox Assess Api Assess Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sandbox Assess",
        "tags": [
          "sandbox"
        ]
      }
    },
    "/api/assessments/{assessment_id}": {
      "get": {
        "operationId": "get_assessment_api_assessments__assessment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "title": "Assessment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-organization-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Assessment",
        "tags": [
          "assessments"
        ]
      }
    },
...
    "/api/factories/{factory_id}/assessments": {
      "get": {
        "operationId": "list_assessments_api_factories__factory_id__assessments_get",
        "parameters": [
          {
            "in": "path",
            "name": "factory_id",
            "required": true,
            "schema": {
              "title": "Factory Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-organization-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AssessmentSummary"
                  },
                  "title": "Response List Assessments Api Factories  Factory Id  Assessments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Assessments",
        "tags": [
          "assessments"
        ]
      },
      "post": {
        "operationId": "create_assessment_api_factories__factory_id__assessments_post",
        "parameters": [
          {
            "in": "path",
            "name": "factory_id",
            "required": true,
            "schema": {
              "title": "Factory Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-organization-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunAssessmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Assessment",
        "tags": [
          "assessments"
        ]
      }
    },
...
    "/api/factories/{factory_id}/scenarios": {
      "get": {
        "operationId": "list_scenarios_api_factories__factory_id__scenarios_get",
        "parameters": [
          {
            "in": "path",
            "name": "factory_id",
            "required": true,
            "schema": {
              "title": "Factory Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-organization-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScenarioOut"
                  },
                  "title": "Response List Scenarios Api Factories  Factory Id  Scenarios Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Scenarios",
        "tags": [
          "assessments"
        ]
      },
      "post": {
        "operationId": "create_scenario_api_factories__factory_id__scenarios_post",
        "parameters": [
          {
            "in": "path",
            "name": "factory_id",
            "required": true,
            "schema": {
              "title": "Factory Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-organization-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Scenario",
        "tags": [
          "assessments"
        ]
      }
    },
...
    "/api/scenarios/{scenario_id}/run": {
      "post": {
        "description": "Apply the scenario's modifications and rerun the same engine.\n\nThe baseline is untouched. The scenario produces its own assessment row with\n`is_baseline = False`, which also keeps scenario runs out of the peer\nbenchmark corpus - a factory's hypothetical must not move its sector.",
        "operationId": "run_scenario_api_scenarios__scenario_id__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-organization-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioComparison"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Scenario",
        "tags": [
          "assessments"
        ]
      }
    },
...
  }
}